% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even_function.R
\name{even_function}
\alias{even_function}
\title{An even function:libr
This function return even numbers from a list.}
\usage{
even_function(list)
}
\arguments{
\item{list}{A list of number}
}
\description{
An even function:libr
This function return even numbers from a list.
}
\examples{
even_function(c(1:10))
even_function(list)
}
\keyword{even}
